Extract args tests
