+++ test/core_test.py
@@ -2,7 +2,6 @@ import unittest
 from collections import namedtuple
 
 from parse_this import create_parser, parse_class
-from parse_this.args import _get_args_and_defaults, _get_args_to_parse
 from parse_this.call import (
     _call,
     _call_method_from_namespace,
@@ -154,28 +153,6 @@ def has_bool_arguments(a: bool):
 
 
 class TestCore(unittest.TestCase):
-    def test_get_args_and_defaults_fill_no_default(self):
-        args_and_defaults = _get_args_and_defaults(
-            ["first", "second", "third"], ("default_value",)
-        )
-        self.assertListEqual(
-            args_and_defaults,
-            [
-                ("first", _NO_DEFAULT),
-                ("second", _NO_DEFAULT),
-                ("third", "default_value"),
-            ],
-        )
-
-    def test_get_args_and_defaults_no_args(self):
-        self.assertListEqual(_get_args_and_defaults([], ()), [])
-
-    def test_get_args_and_defaults_no_default(self):
-        self.assertListEqual(
-            _get_args_and_defaults(["first", "second"], ()),
-            [("first", _NO_DEFAULT), ("second", _NO_DEFAULT)],
-        )
-
     def test_get_default_help_message_no_docstring(self):
         (description, _) = _get_default_help_message(no_docstring, [])
         self.assertIsNotNone(description)
@@ -347,23 +324,6 @@ class TestCore(unittest.TestCase):
             self.assertRaises(
                 SystemExit, parser.parse_args, "yes i_should_be_an_int".split()
             )
-
-    def test_get_args_to_parse_nothing_to_parse(self):
-        self.assertListEqual(_get_args_to_parse(None, []), [])
-
-    def test_get_args_to_parse_remove_prog_from_sys_argv(self):
-        self.assertListEqual(
-            _get_args_to_parse(None, ["prog", "arg_1", "arg_2"]), ["arg_1", "arg_2"]
-        )
-
-    def test_get_args_to_parse_with_options(self):
-        self.assertListEqual(
-            _get_args_to_parse(None, ["prog", "arg", "--kwargs=12"]),
-            ["arg", "--kwargs=12"],
-        )
-
-    def test_get_args_to_parse_used_empty_args_not_sys_argv(self):
-        self.assertListEqual(_get_args_to_parse([], ["prog", "arg_1", "arg_2"]), [])
 
     def test_check_types_not_enough_types_provided(self):
         self.assertRaises(
